SGG-Annotate/requirements.txt
# Tkinter Requirements for SGDET-Annotate COCO Relations Tool
Pillow>=9.0.0
